Make voice in README neutral
The README in the original @aengelberg/loco is amazing in my opinion. It is also quite personal in a couple of sections. I want to adapt these paragraphs in a respectful way with this PR, correcting texts where circumstances have changed due to the fork, while preserving the history and soul of the original project.

I think it's very important that READMEs and documentation in general is up-to-date, especially when the goal is Open Source collaboration, and that's why this makes PR #1.